requirements.txt
crewai==0.5.0
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
sqlalchemy==1.4.49 